Corvo versão 03/07/2024
Ultima versão corvo de acordo com a apresentação final do projeto.

+++ controle-atividades.php
@@ -1,17 +1,17 @@
 <?php
 
-require_once "config.php";
+include "session.php";
 
-if (!isset($_GET["turma"])) {
-    echo "Turma não informada.";
+if(!isset($_GET['turma'])) {
+    echo 'Turma não informada.';
     exit;
 }
 
-$stmt = $conn->prepare("SELECT * FROM corvo_cursos JOIN corvo_turmas ON corvo_turmas.curso = corvo_cursos.id WHERE corvo_turmas.id = :turma");
+$stmt = $conn->prepare("SELECT *, corvo_turmas.id AS idTurma FROM corvo_turmas JOIN corvo_cursos ON corvo_turmas.siglaCurso = corvo_cursos.siglaCurso WHERE corvo_turmas.id = :turma");
 $stmt->bindParam(":turma", $_GET["turma"]);
 $stmt->execute();
 
-$turma = $stmt->fetch(PDO::FETCH_ASSOC);
+$turma = $stmt->fetch();
 
 if (!$turma) {
     echo "Turma não encontrada.";
@@ -19,7 +19,7 @@ if (!$turma) {
 }
 
 
-$stmt = $conn->prepare("SELECT * FROM corvo_usuarios WHERE id = :id");
+$stmt = $conn->prepare("SELECT * FROM corvo_usuarios WHERE matricula = :id");
 $stmt->bindParam(":id", $turma["professor"]);
 $stmt->execute();
 
@@ -29,6 +29,9 @@ if (!$professor) {
     echo "Professor não encontrado.";
     exit;
 }
+
+
+$verificarProfessor = ($professor["matricula"] == $_SESSION['matricula']) ? true : false;
 
 ?>
 <!DOCTYPE html>
@@ -50,16 +53,22 @@ if (!$professor) {
                 <h1 class="h3 mb-3">Atividades</h1>
                 <div>
                     <a href="javascript:history.go(-1)" class="btn btn-outline-primary mx-1">Voltar</a>
+
+                    <?php if($verificarProfessor) {?>
                     <a href="<?= $link ?>/turmas/<?= $_GET["turma"] ?>/atividades/criar" class="btn btn-success"><i class="fa-solid fa-circle-plus"></i></a>
+                    <?php } ?>
                 </div>
             </div>
+            <hr class="my-3">
             <table class="table table-hover table-bordered">
                 <thead>
                     <tr>
                         <th>Atividade</th>
-                        <th>Descrição</th>
                         <th>Data de entrega</th>
+                        <?php if($verificarProfessor) { ?>
+                        <th>Descrição</th>
                         <th>Nota máxima</th>
+                        <?php } ?>
                         <th>Ações</th>
                     </tr>
                 <tbody>
@@ -77,17 +86,21 @@ if (!$professor) {
                         foreach ($atividades as $atividade) {
                             echo "<tr>";
                             echo "<td>" . $atividade["atividade"] . "</td>";
-                            echo "<td>" . ($atividade["descricao"] ? $atividade["descricao"] : "-") . "</td>";
                             echo "<td>" . ($atividade["data_entrega"] ? date("d/m/Y H:i", strtotime($atividade['data_entrega'])) : "Sem data de entrega") . "</td>";
+                            if($verificarProfessor) {
+                            echo "<td>" . ($atividade["descricao"] ? $atividade["descricao"] : "-") . "</td>";
                             echo "<td>" . $atividade["nota_maxima"] . "</td>";
+                            }
                             echo "<td>";
-                            echo "<a href='". $link . "/turmas/" . $_GET["turma"] . "/atividades/" . $atividade["id"] . "/ver' class='btn btn-primary mx-2'><i class='fa-solid fa-eye'></i></a>";
-                            echo "<a href='" . $link . "/turmas/" . $_GET["turma"] . "/atividades/" . $atividade["id"] . "/editar' class='btn btn-warning mx-2'><i class='fa-solid fa-pen-to-square'></i></a>";
-                            echo "<a href='" . $link . "/turmas/" . $_GET["turma"] . "/atividades/" . $atividade["id"] . "/excluir' class='btn btn-danger mx-2'><i class='fa-solid fa-trash'></i></a>";
-
-                            echo "</td>";
+                                echo "<a href='". $link . "/turmas/" . $_GET["turma"] . "/atividades/" . $atividade["id"] . "/ver' class='btn btn-primary mx-2'><i class='fa-solid fa-eye'></i></a>";
+                                if($verificarProfessor) {
+                                echo "<a href='" . $link . "/turmas/" . $_GET["turma"] . "/atividades/" . $atividade["id"] . "/editar' class='btn btn-warning mx-2'><i class='fa-solid fa-pen-to-square'></i></a>";
+                                echo "<a href='" . $link . "/turmas/" . $_GET["turma"] . "/atividades/" . $atividade["id"] . "/excluir' class='btn btn-danger mx-2'><i class='fa-solid fa-trash'></i></a>";
+                                }
+                                
+                                echo "</td>";
+                            } 
                             echo "</tr>";
-                        }
 
 
                     ?>
